+++ examples/example2.cpp
@@ -6,14 +6,14 @@
 
 constexpr double pi = 3.14159265358979323846;
 
-std::vector <float> time0;
-std::vector <float> position0;
-std::vector <std::vector <float>> solutions0;
+std::vector <double> time0;
+std::vector <double> position0;
+std::vector <std::vector <double>> solutions0;
 
-const float p0 = 0.9;
+const double p0 = 0.9;
 
-float f(float x){ return p0 * (std::cos(2*x*pi) - 1);}
-float g(float x){ return 0;}
+double f(double x){ return p0 * (std::cos(2*x*pi) - 1);}
+double g(double x){ return 0;}
 
 int main(){
     const HyperbolicSolver example2 ( f, g, 1, 1, 1, 100, 200);
